Split waybar config into multiple files for easier upgrades
Currently, upgrading omarchy will cause waybar to be clobbered if there
are any changes. Most of these changes are at the module level, so this
probably doesn't need to happen.

All of the waybar config has been moved out into the `default`
directory, which is included in the config. For convenience, the
`modules-` directives are still present in the users local config.

This will need a migration, but after that, we won't need to touch the
users config file unless we add a new module.