feat: improve screen lock security
Hypridle comes with a hypridle.service that automatically
restarts the process on error. We make use of that to ensure
that the screen can always be locked, even if hypridle dies.

To ensure that the screen will always be locked on suspend
we keep hypridle running at all times and inhibit the timers
using a systemd service.