Update uwsm version in the hope of fixing an issue
While working, uwsm will try to generate drop in folder and try to remove a unit called "wayland-wm@.service/timeout.conf". The problem is for some reason this version of uwsm can sometimes mess with the regex (see how it works down below)
unit_ext = re.compile(
        r"[a-zA-Z0-9_:.\\-]+@?\.(service|slice|scope|target|socket|d/[a-zA-Z0-9_:.\\-]+.conf)\2",
        re.MULTILINE,
    )

After some searching, the problem comes from the "\2" at the end which makes wayland-wm@.service/timeout.conf not detected as correct. So as uwsm will just raise an error and the seamless login will just try to restart the process, we will be stuck in an infinite loop of uwsm crashing for this regex and systemd restarting seamless login.