do not fail update if file does not exist
Update will fail if file being deleted does not exist. In my case I had
to create file to proceed with the update. Not sure why it disappeared
from the system as I do not recall deleting it. Anyways trying to delete
a missing file should not fail the update.