Update default hypr/bindings.conf, add omarchy-cmd-cwd
This change adds a working directory parameter to the Alacritty key binding. Now, when you open a new Alacritty instance from an existing one, it will start in the current working directory instead of defaulting to your home directory.

To do this cleanly, we are also adding a new tool called omarchy-cmd-cwd
which returns the current working directory of the active window.