Added script to fix the touchpad on the Rog Flow Z13
Adds install script to fix the touchpad so two finger scrolling
works as well as touch gestures. Before this is just sees the
device as a pointer and not a touchpad.

Fix comes from here:
https://forum.level1techs.com/t/flow-z13-asus-setup-on-linux-may-2025-wip/229551

I tried some other methods to avoid the reload script but this is the
only thing that worked for me. Tested by disabling the service and
removing the files on my local machine and running the script but I
didn't do a full reinstall on the system.