Enable and start Ollama as part of its installation
If Ollama has been installed but not manually started, the following
error will be shown when running `ollama run ...`:

```
Error: ollama server not responding - could not connect to ollama server, run 'ollama serve' to start it
```

Rather than manually running `ollama serve` each time, this ensures that
the service is automatically installed upon installation and restarts,
significally improving the usage experience.

The `bin/omarchy-install-ollama` file was extracted to keep the menu
logic as simple as possible.