first-run: Use balanced profile for desktop as well
My cpu (9550x3d) idles at 30w with the balanced governor, but 45w when in
performance.

This patch changes the default to balanced for both because desktop cpus
also benefit from scaling in the form of power savings (i.e. in my case
this could be $15/year assuming a 80% idle scenario), lower part
degradation, lower heat generation and fan noise.

The CPU will still reach higher clocks when under load, but we still
reap the benefits when on idle.