openai-codex-bin -> openai-codex in menus (#3894)
This was done in a migration in fec2b6a9c4bd7e771308fc0d3b4fc727c857b5e5 but the menu entry was never updated. So, installs after the migration will get the outdated version.

Added a new migration as well for those that installed after the
previous one.