Make `git gui` command work
The `tk` package is flagged as an optional dependency of `git` in Arch's package
manager, but without it the `git gui` command falls over.

I suspect it's optional by default as `git gui` makes no sense in a headless
install, but it's really handy in headmore systems.

This adds 26MB to the installation.

Fixes #3927